move bitlinear layers to separate file
